Remove redundant utils.deepMerge, move utils unit tests
