+++ test/spec/utils_spec.js
@@ -2,7 +2,7 @@ import { getAdServerTargeting } from 'test/fixtures/fixtures.js';
 import { expect } from 'chai';
 import CONSTANTS from 'src/constants.json';
 import * as utils from 'src/utils.js';
-import {waitForElementToLoad} from 'src/utils.js';
+import {cache, memoize1, waitForElementToLoad} from 'src/utils.js';
 
 var assert = require('assert');
 
@@ -1235,5 +1235,55 @@ describe('Utils', function () {
         })
       });
     });
+  });
+
+  describe('memoize', () => {
+    let ncalls;
+
+    function countCalls(fun) {
+      return function (...args) {
+        ncalls++;
+        return fun(...args)
+      }
+    }
+
+    beforeEach(() => {
+      ncalls = 0;
+    });
+
+    describe('on 1 arg', () => {
+      it('delegates to wrapped function', () => {
+        let fun = memoize1((arg) => arg);
+        expect(fun('value')).to.equal('value');
+      })
+
+      it('does not call delegate multiple times for the same argument', () => {
+        let fun = memoize1(countCalls((arg) => arg));
+        fun('value');
+        expect(fun('value')).to.equal('value');
+        expect(ncalls).to.equal(1);
+      });
+
+      it('delegates on cache misses', () => {
+        let fun = memoize1(countCalls((arg) => arg));
+        fun('value0');
+        expect(fun('value1')).to.equal('value1');
+        expect(ncalls).to.equal(2);
+      });
+    });
+
+    describe('on zero args', () => {
+      it('delegates to wrapped function', () => {
+        let fun = cache(() => 'result');
+        expect(fun()).to.equal('result');
+      });
+
+      it('does not call delegate more than once', () => {
+        let fun = cache(countCalls(() => 'result'));
+        fun();
+        expect(fun()).to.equal('result');
+        expect(ncalls).to.equal(1);
+      });
+    })
   });
 });
